requirements.txt
attrs==20.3.0
autopep8==1.5.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.0
Django==2.2
django-allauth==0.44.0
django-storages==1.7.1
djangorestframework==3.12.4
google-api-core==1.26.3
google-auth==1.28.0
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.10.0
iniconfig==1.1.1
oauthlib==3.1.0
packaging==20.9
pluggy==0.13.1
protobuf==3.15.7
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-dotenv==0.17.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.1
